Add files for more days

+++ Day10/day_10.c
@@ -4,10 +4,10 @@
 
 #include "day_10.h"
 
-int day10_1(char input[]) {
+int64_t day10_1(char input[]) {
     return -1;
 }
 
-int day10_2(char input[]) {
+int64_t day10_2(char input[]) {
     return -1;
 }
